+++ packages/bundler/src/validator.ts
@@ -49,7 +49,7 @@ export class OperationValidator {
 
     // Ensure no overlap in given operation
     console.log("Checking in-op conflicts");
-    operation.joinSplitTxs.forEach(({ nullifierA, nullifierB }) => {
+    operation.joinSplits.forEach(({ nullifierA, nullifierB }) => {
       if (opNfSet.has(nullifierA)) {
         return `Conflicting nullifier in operation: ${nullifierA}`;
       }
@@ -80,7 +80,7 @@ export class OperationValidator {
 
     const opDigest = calculateOperationDigest(operation);
     console.log("with digest", opDigest);
-    console.log("with joinsplits", operation.joinSplitTxs);
+    console.log("with joinsplits", operation.joinSplits);
 
     const id = opDigest.toString();
     const bundle: Bundle = { operations: [operation] };