+++ packages/contracts/contracts/CommitmentTreeManager.sol
@@ -41,7 +41,7 @@ contract CommitmentTreeManager is Initializable {
         uint256 indexed oldNoteBNullifier,
         uint128 newNoteAIndex,
         uint128 newNoteBIndex,
-        JoinSplitTransaction joinSplitTx
+        JoinSplit joinSplit
     );
 
     event InsertNoteCommitments(uint256[] commitments);
@@ -66,45 +66,45 @@ contract CommitmentTreeManager is Initializable {
       used as soon as they are checked to be valid.
     */
     function _handleJoinSplit(
-        JoinSplitTransaction calldata joinSplitTx
+        JoinSplit calldata joinSplit
     ) internal {
         // Check validity of both nullifiers
         require(
-            _pastRoots[joinSplitTx.commitmentTreeRoot],
+            _pastRoots[joinSplit.commitmentTreeRoot],
             "Tree root not past root"
         );
         require(
-            !_nullifierSet[joinSplitTx.nullifierA],
+            !_nullifierSet[joinSplit.nullifierA],
             "Nullifier A already used"
         );
         require(
-            !_nullifierSet[joinSplitTx.nullifierB],
+            !_nullifierSet[joinSplit.nullifierB],
             "Nullifier B already used"
         );
         require(
-            joinSplitTx.nullifierA != joinSplitTx.nullifierB,
+            joinSplit.nullifierA != joinSplit.nullifierB,
             "2 nfs should !equal."
         );
 
         // Mark nullifiers as used
-        _nullifierSet[joinSplitTx.nullifierA] = true;
-        _nullifierSet[joinSplitTx.nullifierB] = true;
+        _nullifierSet[joinSplit.nullifierA] = true;
+        _nullifierSet[joinSplit.nullifierB] = true;
 
         // Compute newNote indices in the merkle tree
         uint128 newNoteIndexA = _merkle.getTotalCount();
         uint128 newNoteIndexB = newNoteIndexA + 1;
 
         uint256[] memory noteCommitments = new uint256[](2);
-        noteCommitments[0] = joinSplitTx.newNoteACommitment;
-        noteCommitments[1] = joinSplitTx.newNoteBCommitment;
+        noteCommitments[0] = joinSplit.newNoteACommitment;
+        noteCommitments[1] = joinSplit.newNoteBCommitment;
         insertNoteCommitments(noteCommitments);
 
         emit JoinSplit(
-            joinSplitTx.nullifierA,
-            joinSplitTx.nullifierB,
+            joinSplit.nullifierA,
+            joinSplit.nullifierB,
             newNoteIndexA,
             newNoteIndexB,
-            joinSplitTx
+            joinSplit
         );
     }